% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets}
\alias{assets}
\alias{fly_symbol}
\alias{fly_symbols}
\alias{fly_future_symbol}
\alias{fly_set_symbol_lookup_date}
\alias{fly_sid}
\title{Lookup Equities}
\usage{
fly_symbol(symbol)

fly_symbols(symbol_list)

fly_future_symbol(symbol)

fly_set_symbol_lookup_date(dt)

fly_sid(sid)
}
\arguments{
\item{symbol}{The string ticker symbol for the equity to lookup.}

\item{symbol_list}{The ticker symbols to lookup. Passed as a list like \code{list("AAPL", "MSFT")}.}

\item{dt}{The new symbol lookup date. A \code{Date} or \code{POSIXct} object.}

\item{sid}{The unique integer that identifies an asset.}
}
\description{
Lookup an equity for use in a flyingfox algorithm. These are often assigned
to variables in \code{context} and are used in \code{handle_data()} to perform computations
on the equity's historical prices, or to make an order.
}
\details{
\itemize{
\item \code{fly_symbol()} - Lookup an Equity by its ticker symbol.
\item \code{fly_symbols()} - Lookup multiple Equities as a list.
\item \code{fly_future_symbol()} - Lookup a futures contract with a given symbol.
\item \code{fly_set_symbol_lookup_date()} - Set the date for which symbols will be
resolved to their assets (symbols may map to different firms or underlying
assets at different times)
\item \code{fly_sid()} - Lookup an Asset by its unique asset identifier.
}
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.symbol}{fly_symbol()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.symbols}{fly_symbols()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.future_symbol}{fly_future_symbol()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_symbol_lookup_date}{fly_set_symbol_lookup_date()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.sid}{fly_sid()}
}
}

