% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission.R
\name{commission}
\alias{commission}
\alias{fly_set_commission}
\alias{PerShare}
\alias{PerTrade}
\alias{PerDollar}
\title{Commission}
\usage{
fly_set_commission(us_equities = NULL, us_futures = NULL)

PerShare(cost = 0.001, min_trade_cost = 0)

PerTrade(cost = 0)

PerDollar(cost = 0.0015)
}
\arguments{
\item{us_equities}{The commission model to use for trading US equities}

\item{us_futures}{The commission model to use for trading US futures}

\item{cost}{The amount of commissions paid per dollar/share/trade.}

\item{min_trade_cost}{The minimum amount of commissions paid per trade.}
}
\description{
Control the commission model used for the backtest. Commission
can be calculated per trade, per share, or per dollar.
}
\details{
\itemize{
\item \code{fly_set_commission()} - Set the commission model for the backtest.
\item \code{PerShare()} - Calculates a commission for a transaction based on a per
share cost with an optional minimum cost per trade.
\item \code{PerTrade()} - Calculates a commission for a transaction based on a per
trade cost.
\item \code{PerDollar()} - Calculates a commission for a transaction based on a
per dollar cost.
}
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.finance.commission.CommissionModel}{fly_set_commission()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.commission.PerShare}{PerShare()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.commission.PerTrade}{PerTrade()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.commission.PerDollar}{PerDollar()}
}
}

