% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order-cancellation-policies}
\alias{order-cancellation-policies}
\alias{EODCancel}
\alias{NeverCancel}
\title{Order cancellation policies}
\usage{
EODCancel(warn_on_cancel = TRUE)

NeverCancel()
}
\arguments{
\item{warn_on_cancel}{A logical. Should a warning be raised if this causes an
order to be cancelled?}
}
\description{
For use with \code{fly_set_cancel_policy()}.
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.EODCancel}{EODCancel()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.NeverCancel}{NeverCancel()}
}
}

\seealso{
\code{\link[=fly_set_cancel_policy]{fly_set_cancel_policy()}}
}
