% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order-cancellation}
\alias{order-cancellation}
\alias{fly_set_cancel_policy}
\title{Order cancellation}
\usage{
fly_set_cancel_policy(cancel_policy)
}
\arguments{
\item{cancel_policy}{The cancellation policy to use. Either \code{EODCancel()} or
\code{NeverCancel()}.}
}
\description{
\code{fly_set_cancel_policy()} sets the cancellation policy for the entire
simulation.
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_cancel_policy}{fly_set_cancel_policy()}
}
}

\seealso{
\code{\link[=EODCancel]{EODCancel()}}, \code{\link[=NeverCancel]{NeverCancel()}}
}
