% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order-execution-style}
\alias{order-execution-style}
\alias{MarketOrder}
\alias{LimitOrder}
\alias{StopOrder}
\alias{StopLimitOrder}
\title{Order execution styles}
\usage{
MarketOrder(exchange = NULL)

LimitOrder(limit_price, exchange = NULL)

StopOrder(stop_price, exchange = NULL)

StopLimitOrder(limit_price, stop_price, exchange = NULL)
}
\arguments{
\item{exchange}{The exchange to execute the order on.}

\item{limit_price}{The limit price for the order.}

\item{stop_price}{The stop price for the order.}
}
\description{
The execution style of the order to be placed. This can be a market,
limit, stop or stop-limit order. For use with \code{fly_order()}.
}
\details{
\itemize{
\item \code{MarketOrder()} - Class encapsulating an order to be placed at the current market price.
\item \code{LimitOrder()} - Execution style representing an order to be executed at a
price equal to or better than a specified limit price.
\item \code{StopOrder()} - Execution style representing an order to be placed once the
market price reaches a specified stop price.
\item \code{StopLimitOrder()} - Execution style representing a limit order to be
placed with a specified limit price once the market reaches a specified
stop price.
}
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.finance.execution.MarketOrder}{MarketOrder()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.execution.LimitOrder}{LimitOrder()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.execution.StopOrder}{StopOrder()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.execution.StopOrder}{StopLimitOrder()}
}
}

\seealso{
\code{\link[=fly_order]{fly_order()}}
}
