% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order-utilities}
\alias{order-utilities}
\alias{fly_get_order}
\alias{fly_get_open_orders}
\alias{fly_cancel_order}
\title{Order utilities}
\usage{
fly_get_order(order_id)

fly_get_open_orders(asset = NULL)

fly_cancel_order(order_param)
}
\arguments{
\item{order_id}{The unique string identifier for the order.}

\item{asset}{An asset object. If passed and not \code{NULL}, return only the open orders for the given
asset instead of all open orders.}

\item{order_param}{The \code{order_id} or order object to cancel.}
}
\description{
Get and cancel orders.
}
\details{
\itemize{
\item \code{fly_get_order()} - Lookup an order based on the order id returned from one of the order functions.
\item \code{fly_get_open_orders()} - Retrieve all of the current open orders. If an asset
is passed, this will only return open orders for that asset.
\item \code{fly_cancel_order()} - Cancel an open order.
}
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.get_order}{fly_get_order()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.get_open_orders}{fly_get_open_orders()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.cancel_order}{fly_cancel_order()}
}
}

