% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_to_r}
\alias{py_to_r}
\title{Convert a Python object to R}
\description{
This function is re-exported from the reticulate package.
See \code{\link[reticulate:py_to_r]{reticulate::py_to_r()}} for more details.
}
\seealso{
\code{\link[reticulate:py_to_r]{reticulate::py_to_r()}}
}
\keyword{internal}
