% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-parameters.R
\name{simulation-parameters}
\alias{simulation-parameters}
\alias{fly_set_benchmark}
\title{Set the benchmark asset}
\usage{
fly_set_benchmark(benchmark)
}
\arguments{
\item{benchmark}{An asset. The asset to be used as the benchmark in the
simulation.}
}
\description{
Which asset should be used as the benchmark? Any dividends payed out for that
benchmark asset will be automatically reinvested.
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_benchmark}{fly_set_benchmark()}
}
}

