% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippage.R
\name{slippage}
\alias{slippage}
\alias{fly_set_slippage}
\alias{FixedSlippage}
\alias{VolumeShareSlippage}
\title{Slippage}
\usage{
fly_set_slippage(us_equities = NULL, us_futures = NULL)

FixedSlippage(spread = 0)

VolumeShareSlippage(volume_limit = 0.025, price_impact = 0.1)
}
\arguments{
\item{us_equities}{The slippage model to use for trading US equities}

\item{us_futures}{The slippage model to use for trading US futures}

\item{spread}{\code{spread / 2} will be added to buys and subtracted from sells.}

\item{volume_limit}{The equity volume slippage limit.}

\item{price_impact}{The amount of impact the slippage has on price.}
}
\description{
Control the slippage model used for the backtest. Slippage
can be calculated as a fixed spread or as a function of the volume traded.
}
\details{
\itemize{
\item \code{fly_set_slippage()} - Set the slippage model for the backtest.
\item \code{FixedSlippage()} - Model slippage as a fixed spread.
\item \code{VolumeShareSlippage()} - Model slippage as a function of the volume of contracts traded.
}
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_slippage}{fly_set_slippage()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.slippage.FixedSlippage}{FixedSlippage()}
\item \href{https://www.zipline.io/appendix.html#zipline.finance.slippage.VolumeShareSlippage}{VolumeShareSlippage()}
}
}

