% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading-controls.R
\name{trading-controls}
\alias{trading-controls}
\alias{fly_control_set_do_not_order_list}
\alias{fly_control_set_long_only}
\alias{fly_control_set_max_leverage}
\alias{fly_control_set_max_order_count}
\alias{fly_control_set_max_order_size}
\alias{fly_control_set_max_position_size}
\title{Trading controls}
\usage{
fly_control_set_do_not_order_list(restricted_list, on_error = "fail")

fly_control_set_long_only(on_error = "fail")

fly_control_set_max_leverage(max_leverage)

fly_control_set_max_order_count(max_count, on_error = "fail")

fly_control_set_max_order_size(asset = NULL, max_shares = NULL,
  max_notional = NULL, on_error = "fail")

fly_control_set_max_position_size(asset = NULL, max_shares = NULL,
  max_notional = NULL, on_error = "fail")
}
\arguments{
\item{restricted_list}{A list of assets that cannot be ordered.}

\item{on_error}{How to proceed on a control error. The default is
to stop the backtest. Another option is \code{"log"} to log the error
and continue.}

\item{max_count}{An integer. The maximum number of orders that can be placed
on a single day.}

\item{asset}{An asset. If provided, guard only on positions in
the given asset.}

\item{max_shares}{An integer. The maximum number of shares that can
be ordered at any one time.}

\item{max_notional}{The maximum value that can be ordered at any one time.}
}
\description{
These controls ensure that your algorithm is performing as expected.
The controls act as guards against certain actions such as ordering
black listed stocks, shorting, or exceeding a maximum leverage
threshold.
}
\details{
\itemize{
\item \code{fly_control_set_do_not_order_list()} - Place a restriction on which assets
can be ordered.
\item \code{fly_control_set_long_only()} - Specify that this algorithm cannot take
short positions.
\item \code{fly_control_set_max_leverage()} - Set a limit on the maximum leverage
allowed.
\item \code{fly_control_set_max_order_count()} - Set a limit on the number of orders
that can be placed in a single day.
\item \code{fly_control_set_max_order_size()} - Set a limit on the number of shares
and/or the dollar value of any single order placed for the specified
asset. Limits are treated as absolute values. Limits are only enforced at
the time of order, meaning it is possible to end up with more than the
max number of shares due to splits/dividends, and more than the max
notional due to price improvement.
\item \code{fly_control_set_max_position_size()} - Set a limit on the number of shares
and/or the dollar value held for a single asset.
}
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_do_not_order_list}{fly_control_set_do_not_order_list()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_long_only}{fly_control_set_long_only()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_max_leverage}{fly_control_set_max_leverage()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_max_order_count}{fly_control_set_max_order_count()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_max_order_size}{fly_control_set_max_order_size()}
\item \href{https://www.zipline.io/appendix.html#zipline.api.set_max_position_size}{fly_control_set_max_position_size()}
}
}

