% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrappers.R
\name{zipline-data}
\alias{zipline-data}
\alias{fly_data_history}
\alias{fly_data_current}
\alias{fly_data_is_stale}
\alias{fly_data_can_trade}
\title{Methods for the zipline \code{data} object}
\usage{
fly_data_history(data, assets, fields = "price", bar_count = 100L,
  frequency = "1d")

fly_data_current(data, assets, fields = "price")

fly_data_is_stale(data, assets)

fly_data_can_trade(data, assets)
}
\arguments{
\item{data}{The \code{data} object.}

\item{assets}{A zipline asset or list of assets. Often created from \code{fly_symbol()}.}

\item{fields}{A string field or vector of fields. Accepted fields are
\code{"low"}, \code{"close"}, \code{"volume"}, \code{"price"}, and \code{"last_traded"}.}

\item{bar_count}{Integer number of bars of historical trade data requested.}

\item{frequency}{Frequency of historical data requested.
\code{"1m"} for minutely data or \code{"1d"} for daily data.}
}
\description{
\code{fly_data_history()} is used to retrieve historical data about the asset.
\code{fly_data_current()} is used to retrieve the data as of the current
trading date. See below for additional documentation.
}
\section{Zipline Documentation}{

\itemize{
\item \href{https://www.zipline.io/appendix.html#zipline.protocol.BarData.history}{fly_data_history()}
\item \href{https://www.zipline.io/appendix.html#zipline.protocol.BarData.current}{fly_data_current()}
\item \href{https://www.zipline.io/appendix.html#zipline.protocol.BarData.is_stale}{fly_data_is_stale()}
\item \href{https://www.zipline.io/appendix.html#zipline.protocol.BarData.can_trade}{fly_data_can_trade()}
}
}

